XMLHttpRequest/tests/submissions/Opera/send-redirect-bogus.htm

<!doctype html>
<html>
  <head>
    <title>XMLHttpRequest: send() - Redirects (bogus Location header)</title>
    <script src="/resources/testharness.js"></script>
    <script src="/resources/testharnessreport.js"></script>
  </head>
  <body>
    <div id="log"></div>
    <script>
      function redirect(code, location) {
        var test = async_test(document.title + " (" + code + ": " + location + ")")
        test.step(function() {
          var client = new XMLHttpRequest()
          client.onreadystatechange = function() {
            test.step(function() {
              if(client.readyState == 4) {
                assert_equals(client.status, 0)
                assert_equals(client.statusText, "")
                test.done()
              }
            })
          }
          client.open("GET", "resources/redirect.php?location=" + location + "&code=" + code)
          client.send(null)
        })
      }
      redirect("302", "http://z")
      redirect("302", "mailto:someone@example.org")
      redirect("303", "http://z")
      redirect("303", "mailto:someone@example.org")
    </script>
  </body>
</html>
